+++ tests/devices/eiger/test_eiger_filewriter.py
@@ -1,0 +1,12 @@
+import pytest
+
+from tickit.devices.eiger.filewriter.eiger_filewriter import EigerFileWriter
+
+
+@pytest.fixture
+def filewriter() -> EigerFileWriter:
+    return EigerFileWriter()
+
+
+def test_eiger_filewriter_constructor():
+    EigerFileWriter()