+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added:
 
 - HTTP REST Adapter & example device
 - Example Counter device
+- Basic Eiger device without major logic
 - Unit tests covering:
 
   - EPICS Adapter